+ Register + generate RSA Keys

+++ src/secureApp/server/ServiceChat.java
@@ -62,7 +62,7 @@ public class ServiceChat implements Runnable {
 
     private KeyFactory factory;
 	private PublicKey pub;
-	private PrivateKey priv;
+	//private PrivateKey priv;
     private final int DATASIZE = 128;	
     
 
@@ -71,67 +71,26 @@ public class ServiceChat implements Runnable {
         this.NBMAXUSERCONNECTED = NBMAXUSERCONNECTED;
         this.in = new Scanner(socket.getInputStream());
         this.client = new Client(new PrintWriter(socket.getOutputStream(), true), socket);
-
-        try {
-            generateRSAKeys();
-		} catch( Exception e ) {
-			System.out.println( "initNewCard: " + e );
-		}
     }
 
     public ServiceChat(int NBMAXUSERCONNECTED) { // Admin
         this.NBMAXUSERCONNECTED = NBMAXUSERCONNECTED;
         this.in = new Scanner(System.in);
         this.client = new Client(new PrintWriter(new OutputStreamWriter(System.out, StandardCharsets.UTF_8), true));
-        
-        try {
-            generateRSAKeys();
-		} catch( Exception e ) {
-			System.out.println( "initNewCard: " + e );
-		}
 
         // development purpose
-        loadBdd("/loadbdd users.db");
+        // loadBdd("/loadbdd users.db");
     }
 
-    private void generateRSAKeys() throws Exception {
-        byte[] modulus_b = new byte[] {
-            (byte)0x90,(byte)0x08,(byte)0x15,(byte)0x32,(byte)0xb3,(byte)0x6a,(byte)0x20,(byte)0x2f,
-            (byte)0x40,(byte)0xa7,(byte)0xe8,(byte)0x02,(byte)0xac,(byte)0x5d,(byte)0xec,(byte)0x11,
-            (byte)0x1d,(byte)0xfa,(byte)0xf0,(byte)0x6b,(byte)0x1c,(byte)0xb7,(byte)0xa8,(byte)0x39,
-            (byte)0x19,(byte)0x50,(byte)0x9c,(byte)0x44,(byte)0xed,(byte)0xa9,(byte)0x51,(byte)0x01,
-            (byte)0x0f,(byte)0x11,(byte)0xd6,(byte)0xa3,(byte)0x60,(byte)0xa7,(byte)0x7e,(byte)0x95,
-            (byte)0xa2,(byte)0xfa,(byte)0xe0,(byte)0x8d,(byte)0x62,(byte)0x5b,(byte)0xf2,(byte)0x62,
-            (byte)0xa2,(byte)0x64,(byte)0xfb,(byte)0x39,(byte)0xb0,(byte)0xf0,(byte)0x6f,(byte)0xa2,
-            (byte)0x23,(byte)0xae,(byte)0xbc,(byte)0x5d,(byte)0xd0,(byte)0x1a,(byte)0x68,(byte)0x11,
-            (byte)0xa7,(byte)0xc7,(byte)0x1b,(byte)0xda,(byte)0x17,(byte)0xc7,(byte)0x14,(byte)0xab,
-            (byte)0x25,(byte)0x92,(byte)0xbf,(byte)0xcc,(byte)0x81,(byte)0x65,(byte)0x7a,(byte)0x08,
-            (byte)0x90,(byte)0x59,(byte)0x7f,(byte)0xc4,(byte)0xf9,(byte)0x43,(byte)0x9c,(byte)0xaa,
-            (byte)0xbe,(byte)0xe4,(byte)0xf8,(byte)0xfb,(byte)0x03,(byte)0x74,(byte)0x3d,(byte)0xfb,
-            (byte)0x59,(byte)0x7a,(byte)0x56,(byte)0xa3,(byte)0x19,(byte)0x66,(byte)0x43,(byte)0x77,
-            (byte)0xcc,(byte)0x5a,(byte)0xae,(byte)0x21,(byte)0xf5,(byte)0x20,(byte)0xa1,(byte)0x22,
-            (byte)0x8f,(byte)0x3c,(byte)0xdf,(byte)0xd2,(byte)0x03,(byte)0xe9,(byte)0xc2,(byte)0x38,
-            (byte)0xe7,(byte)0xd9,(byte)0x38,(byte)0xef,(byte)0x35,(byte)0x82,(byte)0x48,(byte)0xb7
-        };
-        byte[] private_exponent_b = new byte[] {
-            (byte)0x69,(byte)0xdf,(byte)0x67,(byte)0x25,(byte)0xa3,(byte)0xb8,(byte)0x88,(byte)0xfb,
-            (byte)0xf2,(byte)0xfc,(byte)0xf9,(byte)0x90,(byte)0xad,(byte)0x7f,(byte)0x44,(byte)0xbd,
-            (byte)0xb8,(byte)0x59,(byte)0xf3,(byte)0x4b,(byte)0xe9,(byte)0x0a,(byte)0x1f,(byte)0x80,
-            (byte)0x09,(byte)0x59,(byte)0xb5,(byte)0xe4,(byte)0xfd,(byte)0x06,(byte)0x0e,(byte)0xe3,
-            (byte)0x46,(byte)0x5e,(byte)0x88,(byte)0x76,(byte)0x03,(byte)0xe0,(byte)0x5b,(byte)0x2e,
-            (byte)0x47,(byte)0x65,(byte)0x3e,(byte)0x96,(byte)0xef,(byte)0x0c,(byte)0x43,(byte)0x79,
-            (byte)0xb9,(byte)0x81,(byte)0x9d,(byte)0x21,(byte)0xe5,(byte)0x2c,(byte)0x78,(byte)0x02,
-            (byte)0xa9,(byte)0x54,(byte)0x12,(byte)0x66,(byte)0xab,(byte)0x48,(byte)0x1d,(byte)0xe2,
-            (byte)0x6e,(byte)0x1d,(byte)0x7d,(byte)0xb2,(byte)0xce,(byte)0x7a,(byte)0x3f,(byte)0xbb,
-            (byte)0x34,(byte)0xf2,(byte)0x46,(byte)0x5f,(byte)0x73,(byte)0x7c,(byte)0xba,(byte)0xf8,
-            (byte)0xc1,(byte)0x29,(byte)0x97,(byte)0x85,(byte)0x67,(byte)0xdf,(byte)0x82,(byte)0x87,
-            (byte)0x89,(byte)0x61,(byte)0x42,(byte)0xcc,(byte)0x1d,(byte)0xcc,(byte)0x03,(byte)0xce,
-            (byte)0x41,(byte)0x7d,(byte)0x8f,(byte)0x25,(byte)0xc1,(byte)0x61,(byte)0xfe,(byte)0x06,
-            (byte)0x4f,(byte)0x1a,(byte)0xf2,(byte)0x48,(byte)0x55,(byte)0xd8,(byte)0x6e,(byte)0xc6,
-            (byte)0x3f,(byte)0x6d,(byte)0xe1,(byte)0xce,(byte)0xa9,(byte)0x28,(byte)0x9e,(byte)0x03,
-            (byte)0x2d,(byte)0x74,(byte)0x59,(byte)0x1c,(byte)0xdb,(byte)0x18,(byte)0xb3,(byte)0x41
-        };
-        byte[] public_exponent_b = new byte[] { (byte)0x01,(byte)0x00,(byte)0x01 };
+    private synchronized void generateRSAKeys(Client client) throws Exception {
+
+        byte[] modulus_b = Base64.getDecoder().decode(client.getRSAModulus());
+        byte[] public_exponent_b = Base64.getDecoder().decode(client.getRSAExponent());
+
+        System.out.println("public modulus : ");
+        displayBytes(modulus_b);
+        System.out.println("public exponent : ");
+        displayBytes(public_exponent_b);
 
         // Transform byte[] into String
 		String mod_s =  HexString.hexify( modulus_b );
@@ -142,24 +101,26 @@ public class ServiceChat implements Runnable {
 		pub_s = pub_s.replaceAll( " ", "" );
 		pub_s = pub_s.replaceAll( "\n", "" );
 
-		String priv_s =  HexString.hexify( private_exponent_b );
-		priv_s = priv_s.replaceAll( " ", "" );
-		priv_s = priv_s.replaceAll( "\n", "" );
-
 		// Load the keys from String into BigIntegers 
 		BigInteger modulus = new BigInteger(mod_s, 16);
 		BigInteger pubExponent = new BigInteger(pub_s, 16);
-		BigInteger privExponent = new BigInteger(priv_s, 16);
 
 		// Create private and public key specs from BinIntegers 
 		RSAPublicKeySpec publicSpec = new RSAPublicKeySpec(modulus, pubExponent);
-		RSAPrivateKeySpec privateSpec = new RSAPrivateKeySpec(modulus, privExponent);
 
 		// Create the RSA private and public keys 
 		this.factory = KeyFactory.getInstance("RSA");
 		this.pub = factory.generatePublic(publicSpec);
-		this.priv = factory.generatePrivate(privateSpec);
     }
+
+    private void displayBytes(byte[] bytes){
+		int i = 0;
+		for (byte b : bytes) {
+			System.out.printf("%02X ", b);
+			if (++i%8 == 0)
+				System.out.println("");
+		}
+	}
 
     private synchronized boolean userConnectedLimitReachedCheck() throws IOException {
         if (ServiceChat.connectedClients.size() >= NBMAXUSERCONNECTED){
@@ -269,13 +230,9 @@ public class ServiceChat implements Runnable {
                 this.client.getOut().close();
                 this.socket.close();
             } else {
-
                 try {
-
-                    /* Get public key section */
-
-
-
+                    /* Generate client RSA Keys */
+                    generateRSAKeys(registeredClients.get(isFound));
 
                     /* Challenge section */
                     byte[] challengeBytes = generateChallenge();
@@ -330,10 +287,11 @@ public class ServiceChat implements Runnable {
 
         if(!username.toLowerCase(Locale.ROOT).equals("admin")) {
             if(!usernameExists(username)){
-                this.client.setUsername(username);
-                registeredClients.add(this.client);
+                cardRegister(username);
+                return false;
+            } else {
+                return cardLogin(username); // false = auth KO; true = auth OK``
             }
-            return cardLogin(username); // false = auth KO; true = auth OK
         } else {
             this.client.setUsername("Forbidden username");
             logout(this.client);
@@ -342,6 +300,32 @@ public class ServiceChat implements Runnable {
     }
 
 
+    private synchronized void cardRegister(String username) throws IOException {
+        this.client.getOut().println("<SYSTEM> Register");
+        this.client.setUsername(username);
+
+        this.client.getOut().println("<SYSTEM> REGISTRATION NEW");
+
+        while (this.in.hasNextLine()) {
+            String raw = this.in.nextLine().trim();
+            if(raw.startsWith("<SYSTEM> REGISTRATION MODULUS")) {
+                System.out.println(raw);
+                this.client.setRSAModulus(raw.split(" ")[3]);
+            } else if (raw.startsWith("<SYSTEM> REGISTRATION EXPONENT")) {
+                System.out.println(raw);
+                this.client.setRSAExponent(raw.split(" ")[3]);
+            }
+            if (this.client.getRSAModulus() != null && this.client.getRSAExponent() != null) {
+                break;
+            } 
+        }
+
+        registeredClients.add(this.client);
+
+        ServerChat.logger.log(Level.INFO, "<SYSTEM> [REGISTER]: Registering " + this.client.getUsername());
+        this.client.getOut().println("<SYSTEM> Registration Successful");
+        this.socket.close();
+    }
 
     private synchronized void register(String username) throws IOException {
         this.client.getOut().println("<SYSTEM> Register");
@@ -447,6 +431,11 @@ public class ServiceChat implements Runnable {
     }
 
     private synchronized void saveBdd(String raw) {
+        if (raw.split(" ").length != 2) {
+            System.out.println("<SYSTEM> Command error");
+            return;
+        }
+
         String bddFile = "../src/secureApp/server/Databases/" + raw.split(" ")[1];
         try {
             new File(bddFile).delete();
@@ -455,16 +444,21 @@ public class ServiceChat implements Runnable {
             int cpt = 0;
             for (Client client: registeredClients) {
                 ++cpt;
-                writer.write(client.getUsername() + ":" + client.getPassword() + "\n");
+                writer.write(client.getUsername() + ":" + client.getRSAExponent() + ":" + client.getRSAModulus() + "\n");
             }
             writer.close();
             ServerChat.logger.log(Level.INFO, "<SYSTEM> [SAVEBDD]: Database saved, saved " + cpt + " new users");
         } catch (IOException e) {
-            e.printStackTrace();
+            e.getMessage();
         }
     }
 
     private synchronized void loadBdd(String raw) {
+        if (raw.split(" ").length != 2) {
+            System.out.println("<SYSTEM> Command error");
+            return;
+        }
+
         String bddFile = "../src/secureApp/server/Databases/" + raw.split(" ")[1];
         try {
             BufferedReader reader = new BufferedReader(new FileReader(bddFile));
@@ -472,14 +466,14 @@ public class ServiceChat implements Runnable {
             while ((line = reader.readLine()) != null) {
                 ++cpt;
                 String[] splitLine = line.split(":");
-                registeredClients.add(new Client(splitLine[0], splitLine[1]));
+                registeredClients.add(new Client(splitLine[0], splitLine[1], splitLine[2]));
             }
             registeredClients = registeredClients.stream().distinct().collect(Collectors.toList());
             reader.close();
             ServerChat.logger.log(Level.INFO, "<SYSTEM> [LOADBDD]: Database loaded, registered " + cpt + " new users");
 
         } catch (IOException e) {
-            e.printStackTrace();
+            e.getMessage();
         }
     }
